Move GPU detection to platform abstraction layer
- Rename gpu_detection.py to hardware_detection.py (thin wrapper)
- All GPU/runtime/CT2 detection and display logic in platform/windows/gpu.py
- macOS gpu.py is a single no-op (no GPU support)
- main.py call simplified to detect_hardware(configured_device)
