+++ src/whisper_key/platform/macos/gpu.py
@@ -1,0 +1,2 @@
+def detect_and_print(configured_device):
+    pass
